dev/requirements.txt
kafka-python
selenium
elasticsearch
configparser
loguru
requests
beautifulsoup4